Enviando exceptions e ajustes no retorno do pedido

+++ eccomerce/src/main/java/org/serratec/com/backend/ecommerce/services/PedidoService.java
@@ -1,24 +1,29 @@
 package org.serratec.com.backend.ecommerce.services;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 import java.util.stream.Collectors;
 
+import org.serratec.com.backend.ecommerce.configs.MailConfig;
+import org.serratec.com.backend.ecommerce.entities.CarrinhoEntity;
 import org.serratec.com.backend.ecommerce.entities.PedidoEntity;
+import org.serratec.com.backend.ecommerce.entities.dto.CadastroPedidoDto;
+import org.serratec.com.backend.ecommerce.entities.dto.CarrinhoDto;
+import org.serratec.com.backend.ecommerce.entities.dto.PedidoDto;
 import org.serratec.com.backend.ecommerce.entities.dto.ProdutoDto;
 import org.serratec.com.backend.ecommerce.entities.dto.ProdutosPedidosDto;
-import org.serratec.com.backend.ecommerce.entities.dto.PedidoDto;
-import org.serratec.com.backend.ecommerce.entities.dto.CarrinhoDto;
 import org.serratec.com.backend.ecommerce.enums.StatusCompra;
 import org.serratec.com.backend.ecommerce.exceptions.CarrinhoException;
 import org.serratec.com.backend.ecommerce.exceptions.EntityNotFoundException;
+import org.serratec.com.backend.ecommerce.exceptions.PedidoException;
 import org.serratec.com.backend.ecommerce.exceptions.ProdutoException;
-import org.serratec.com.backend.ecommerce.mappers.ProdutoMapper;
 import org.serratec.com.backend.ecommerce.mappers.PedidoMapper;
+import org.serratec.com.backend.ecommerce.mappers.ProdutoMapper;
+import org.serratec.com.backend.ecommerce.repositories.CarrinhoRepository;
 import org.serratec.com.backend.ecommerce.repositories.PedidoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -41,6 +46,12 @@ public class PedidoService {
 
 	@Autowired
 	CarrinhoService cartService;
+	
+	@Autowired
+	CarrinhoRepository carrinhoRepository;
+	
+	@Autowired
+	MailConfig mailconfig;
 
 	public PedidoEntity findById(Long id) throws EntityNotFoundException {
 		return repository.findById(id).orElseThrow(() -> new EntityNotFoundException(id + " não encontrado."));
@@ -50,9 +61,9 @@ public class PedidoService {
 		return repository.findAll().stream().map(mapper::toDto).collect(Collectors.toList());
 	}
 
-	public PedidoDto getById(Long id) throws EntityNotFoundException {
-		return mapper.toDto(this.findById(id));
-	}
+//	public PedidoDto getById(Long id) throws EntityNotFoundException {
+//		return mapper.toDto(this.findById(id));
+//	}
 
 	public PedidoDto getByNumeroPedido(String numeroPedido) throws EntityNotFoundException {
 		return mapper.toDto(repository.findByNumeroPedido(numeroPedido));
@@ -69,25 +80,33 @@ public class PedidoService {
 		return repository.save(mapper.toEntity(purchase));
 	}
 
-	public PedidoDto update(Long id, PedidoDto purchaseUpdate) throws EntityNotFoundException {
-		PedidoEntity purchase = this.findById(id);
-		purchase.setNumeroPedido(purchaseUpdate.getNumeroPedido());
-		purchase.setValorTotal(purchaseUpdate.getValorTotal());
-		purchase.setDataPedido(purchaseUpdate.getDataPedido());
-		purchase.setDataEntrega(purchaseUpdate.getDataEntrega());
-		purchase.setStatus(purchaseUpdate.getStatus());
+//	public PedidoDto update(Long id, PedidoDto purchaseUpdate) throws EntityNotFoundException {
+//		PedidoEntity purchase = this.findById(id);
+//		purchase.setNumeroPedido(purchaseUpdate.getNumeroPedido());
+//		purchase.setValorTotal(purchaseUpdate.getValorTotal());
+//		purchase.setDataPedido(purchaseUpdate.getDataPedido());
+//		purchase.setDataEntrega(purchaseUpdate.getDataEntrega());
+//		purchase.setStatus(purchaseUpdate.getStatus());
+//
+//		return mapper.toDto(repository.save(purchase));
+//	}
 
-		return mapper.toDto(repository.save(purchase));
-	}
-
-	public void delete(Long id) throws EntityNotFoundException, DataIntegrityViolationException {
-		try {
-			if (this.findById(id) != null) {
+	public void delete(Long id) throws EntityNotFoundException, PedidoException {
+		if (this.findById(id) != null) {
+			if(this.findById(id).getStatus().equals(StatusCompra.NAO_FINALIZADO)) {
+				List<CarrinhoEntity> carrinho = carrinhoRepository.findByPedidos(this.findById(id));
+				for (CarrinhoEntity carrinhoEntity : carrinho) {
+					if(carrinhoEntity.getPedidos().getId().equals(id)) {
+						carrinhoRepository.delete(carrinhoEntity);
+					}
+				}
 				repository.deleteById(id);
+			}else {
+				throw new PedidoException(
+						"Pedido com id: " + id + " já finalizado");
 			}
-		} catch (DataIntegrityViolationException e) {
-			throw new DataIntegrityViolationException(
-					"Categoria com id: " + id + " está associada a um ou mais produtos, favor verificar");
+		} else {
+			throw new EntityNotFoundException("Pedido com id: " + id + "não encontrado!");
 		}
 
 	}
@@ -101,7 +120,7 @@ public class PedidoService {
 		return numeroPedido;
 	}
 
-	public PedidoDto order(PedidoDto purchase) throws EntityNotFoundException, ProdutoException {
+	public CadastroPedidoDto order(PedidoDto purchase) throws EntityNotFoundException, ProdutoException {
 		Long idPedido = this.create(purchase).getId();
 
 		List<ProdutoDto> produtos = new ArrayList<>();
@@ -133,7 +152,12 @@ public class PedidoService {
 		entity.setValorTotal(cartService.calcularTotal(idPedido));
 
 		repository.save(entity);
-		return mapper.toDto(entity);
+
+		PedidoDto pedido = mapper.toDto(entity);
+		pedido.setProduto(purchase.getProduto());
+
+		return mapper.toCadastroPedidoDto(pedido);
+
 	}
 
 	public PedidoDto updateOrder(String numeroPedido, List<ProdutosPedidosDto> produtosCarrinho)
@@ -191,6 +215,37 @@ public class PedidoService {
 
 		repository.save(entity);
 
-		return mapper.toDto(entity);
+		PedidoDto pedido = mapper.toDto(entity);
+		pedido.setProduto(produtosCarrinho);
+
+		return pedido;
+	}
+
+	public PedidoDto finalizarPedido(String numeroPedido) throws EntityNotFoundException {
+		PedidoEntity entity = repository.findByNumeroPedido(numeroPedido);
+		List<ProdutosPedidosDto> produtos = new ArrayList<>();
+		List<CarrinhoEntity> carrinho = carrinhoRepository.findByPedidos(entity);
+		for (CarrinhoEntity carrinhoEntity : carrinho) {
+			ProdutosPedidosDto produto = productMapper
+					.toProdutosPedidos(productService.findByName(carrinhoEntity.getProdutos().getNome()));
+			produto.setQuantidade(carrinhoEntity.getQuantidade());
+
+			produtos.add(produto);
+		}
+		
+		entity.setDataPedido(LocalDate.now());
+		entity.setDataEntrega(LocalDate.now().plusDays(7));
+		entity.setStatus(StatusCompra.FINALIZADO);
+
+		String msg = "Recebemos seu pedido";
+
+		mailconfig.sendMail(entity.getCliente().getEmail(), "Pedido recebido com sucesso", msg);
+
+		PedidoDto pedido = mapper.toDto(entity);
+		pedido.setProduto(produtos);
+
+		repository.save(entity);
+		return (pedido);
+
 	}
 }
